dump WIP commit of lsdem developments started by Jibril
the `fix_rigid_lsdem` is a copy of fix_rigid_small stripped from all the attributes
that are not useful for LSDEM (Langevin, extended particles etc)

the `lsdem_const` is a copy of `rigid_const.h` where, as of now, only the inertia
constants have been stripped. There remains a lot of likely useless consts in the file

Generally, the current state of these files is very unstable. It likely does not build.
I had started modifications to do the I/O and left it in the middle of work with nothing
consolidated. I remember I had started to code a way to write chunks of lines from the file
but Rob told me one line per entry was used and that might actually simplify things a lot
(at the cost of longer files).

Couple random remarks: the so-called restart file in the documentation is a text file,
not a binary file. I got confused about how to handle it until Dan told me. This will
simplify things but has not been taken care of yet.
I may have left enough comments to guide things but if you need me to help stabilize
these files I'll be there.